+++ python/sdist/amici/importers/petab/v1/__init__.py
@@ -1,4 +1,4 @@
-"""PEtab import related code."""
+"""PEtab v1 import related code."""
 
 # ID of model parameter that is to be added to SBML model to indicate
 #  preequilibration
