+++ python/sdist/amici/importers/petab/v1/parameter_mapping.py
@@ -40,9 +40,9 @@ from petab.v1.models import MODEL_TYPE_PYSB, MODEL_TYPE_SBML
 from sympy.abc import _clash
 
 import amici
+from amici import AmiciModel
 from amici.importers.sbml import get_species_initial
 
-from .. import AmiciModel
 from . import PREEQ_INDICATOR_ID
 from .util import get_states_in_condition_table
