+++ python/sdist/amici/importers/petab/v1/petab_import.py
@@ -16,8 +16,8 @@ import petab.v1 as petab
 from petab.v1.models import MODEL_TYPE_PYSB, MODEL_TYPE_SBML
 
 import amici
+from amici.logging import get_logger
 
-from ..logging import get_logger
 from .import_helpers import (
     _can_import_model,
     _create_model_name,
