+++ tests/petab_test_suite/test_petab_suite.py
@@ -13,10 +13,13 @@ import pytest
 from _pytest.outcomes import Skipped
 from amici import SteadyStateSensitivityMode
 from amici.gradient_check import check_derivatives as amici_check_derivatives
+from amici.importers.petab.v1 import (
+    import_petab_problem,
+    rdatas_to_measurement_df,
+    simulate_petab,
+)
+from amici.importers.petab.v1.conditions import create_parameterized_edatas
 from amici.logging import get_logger, set_log_level
-from amici.petab.conditions import create_parameterized_edatas
-from amici.petab.petab_import import import_petab_problem
-from amici.petab.simulations import rdatas_to_measurement_df, simulate_petab
 
 logger = get_logger(__name__, logging.DEBUG)
 set_log_level(get_logger("amici.petab_import"), logging.DEBUG)
