+++ tests/sciml/test_sciml.py
@@ -13,12 +13,12 @@ import numpy as np
 import pandas as pd
 import petab.v1 as petab
 import pytest
+from amici.importers.petab.v1 import import_petab_problem
 from amici.jax import (
     generate_equinox,
     petab_simulate,
     run_simulations,
 )
-from amici.petab import import_petab_problem
 from petab_sciml import NNModelStandard
 from yaml import safe_load
